fixing errors of matrix caps

+++ tf2_models/caps_util.py
@@ -210,7 +210,7 @@ def to_sparse(probs, spatial_routing_matrix, sparse_filler=tf.math.log(1e-20)):
   batch_idx = tf.reshape(batch_idx, [batch_size, parent_space_2, kk])
 
   # Combine the 3 coordinates
-  indices = tf.stack((batch_idx, parent_idx, child_sparse_idx), axis=3)
+  indices = tf.stack((batch_idx, parent_idx, tf.cast(child_sparse_idx, dtype=tf.int32)), axis=3)
   indices = tf.constant(indices)
 
   # Convert each spatial location to sparse
